@@ -4,9 +4,13 @@
   "description": "Global forest watch - Biomass Loss API",
   "main": "index.js",
   "scripts": {
-    "coverage": "NODE_PATH=app/src node ./node_modules/.bin/grunt --gruntfile app/Gruntfile.js nyc",
-    "start": "NODE_PATH=app/src node app/index.js",
-    "test": "NODE_PATH=app/src NODE_ENV=test grunt --gruntfile app/Gruntfile.js test"
+    "coverage": "nyc ts-mocha -b --project tsconfig.json -r tsconfig-paths/register --timeout 20000 'test/**/*.ts' --exit",
+    "test": "ts-mocha -b --project tsconfig.json -r tsconfig-paths/register --timeout 20000 'test/**/*.ts' --exit",
+    "start": "ts-node --files --project tsconfig.json -r tsconfig-paths/register src/index.ts",
+    "watch": "ts-node-dev --respawn --transpile-only --files --project tsconfig.json -r tsconfig-paths/register src/index.ts",
+    "build": "tsc",
+    "lint": "eslint \"{src,test}/**/*.ts\" --fix",
+    "prepare": "husky install"
   },
   "repository": {
     "type": "git",
@@ -23,7 +27,7 @@
   "author": {
     "name": "Vizzuality",
     "email": "hello@vizzuality.com",
-    "url": "http://vizzuality.com/"
+    "url": "https://vizzuality.com/"
   },
   "license": "MIT",
   "bugs": {
@@ -31,51 +35,48 @@
   },
   "homepage": "https://github.com/gfw-api/gfw-biomass-losss-api#readme",
   "dependencies": {
-    "bunyan": "1.7.1",
-    "cartodb": "0.4.0",
-    "co": "4.6.0",
-    "co-request": "1.0.0",
-    "config": "1.19.0",
-    "geojsonhint": "1.2.0",
-    "jsonapi-serializer": "2.1.1",
-    "koa": "1.1.2",
-    "koa-bodyparser": "2.0.1",
-    "koa-cash": "2.0.0",
-    "koa-convert": "^2.0.0",
-    "koa-logger": "1.3.0",
-    "koa-mount": "1.3.0",
-    "koa-router": "5.4.0",
+    "bunyan": "^1.8.15",
+    "cartodb": "0.5.1",
+    "config": "^3.3.9",
+    "jsonapi-serializer": "^3.6.7",
+    "koa": "^2.14.2",
+    "koa-bodyparser": "^4.4.1",
+    "koa-cash": "^4.1.1",
+    "koa-logger": "^3.2.1",
+    "koa-router": "^12.0.0",
     "koa-simple-healthcheck": "^0.0.1",
-    "koa-validate": "0.2.11",
-    "lru-cache": "4.0.1",
+    "lru-cache": "^10.0.0",
     "mustache": "2.2.1",
-    "python-shell": "0.4.0",
-    "rw-api-microservice-node": "^3.4.1"
+    "rw-api-microservice-node": "^5.1.3",
+    "typescript": "^5.1.6"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
-    "chai-http": "^4.3.0",
-    "co-mocha": "1.1.2",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.1.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-mocha": "^6.3.0",
-    "eslint-plugin-react": "^7.19.0",
-    "grunt": "^1.1.0",
-    "grunt-cli": "^1.3.2",
-    "grunt-contrib-clean": "^2.0.0",
-    "grunt-contrib-jshint": "^2.1.0",
-    "grunt-contrib-watch": "^1.1.0",
-    "grunt-express-server": "^0.5.4",
-    "grunt-mocha-test": "^0.13.3",
-    "grunt-notify": "^0.4.5",
-    "grunt-simple-nyc": "^3.0.1",
-    "husky": "^4.2.3",
-    "lint-staged": "^10.0.8",
-    "load-grunt-tasks": "^5.1.0",
-    "mocha": "^7.1.1",
-    "nock": "^12.0.3",
-    "nyc": "^15.0.0"
+    "@types/bunyan": "^1.8.8",
+    "@types/config": "^3.3.0",
+    "@types/jsonapi-serializer": "^3.6.5",
+    "@types/mustache": "^4.2.2",
+    "@types/koa": "^2.13.8",
+    "@types/koa-cash": "^4.1.0",
+    "@types/koa-bodyparser": "^4.3.10",
+    "@types/koa-logger": "^3.1.2",
+    "@types/koa-router": "^7.4.4",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^20.4.9",
+    "@typescript-eslint/eslint-plugin": "^5.44.0",
+    "@typescript-eslint/parser": "^5.44.0",
+    "chai": "^4.3.7",
+    "chai-datetime": "^1.8.0",
+    "chai-http": "^4.4.0",
+    "eslint": "^8.46.0",
+    "eslint-plugin-mocha": "^10.1.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.2.3",
+    "mocha": "^10.2.0",
+    "nock": "^13.3.2",
+    "nyc": "^15.1.0",
+    "ts-mocha": "^10.0.0",
+    "ts-node": "^10.9.1",
+    "ts-node-dev": "^2.0.0",
+    "tsconfig-paths": "^4.2.0"
   }
 }
