@@ -1,10 +1,11 @@
+class NotFound extends Error {
 
-class NotFound extends Error{
-
-    constructor(message){
+    constructor(message: string) {
         super(message);
         this.name = 'NotFound';
         this.message = message;
     }
+
 }
-module.exports = NotFound;
+
+export default NotFound;
